+++ packages/agents/src/index.ts
@@ -1,5 +1,5 @@
-export * from './core/types';
-export * from './config/agentConfigs';
+export * from './types';
+export * from './config';
 export * from './ragAgentFactory';
 export * from './db/vectorStore';
 export * from './config';
