+++ packages/agents/src/ragAgentFactory.ts
@@ -1,24 +1,20 @@
 import { BaseMessage } from '@langchain/core/messages';
 import { Embeddings } from '@langchain/core/embeddings';
-import {
-  AvailableAgents,
-  getAgentConfig,
-  LLMConfig,
-} from './config/agentConfigs';
+import { getAgentConfig } from './config/agent';
 import EventEmitter from 'events';
 import { RagPipeline } from './pipeline/ragPipeline';
 import { VectorStore } from './db/vectorStore';
+import { LLMConfig } from './types';
 
 export class RagAgentFactory {
   static createAgent(
-    agentName: AvailableAgents,
     message: string,
     history: BaseMessage[],
     llm: LLMConfig,
     embeddings: Embeddings,
     vectorStore: VectorStore,
   ): EventEmitter {
-    const config = getAgentConfig(agentName, vectorStore);
+    const config = getAgentConfig(vectorStore);
     const pipeline = new RagPipeline(llm, embeddings, config);
     return pipeline.execute({
       query: message,
